Require title in posts frontmatter
The fact that getting it from headers is async is just turning out to be
a pain. Everywhere I need the title, having to call those utility
functions. This is leading up to setting up the RSS feed, which also
needs (ideally _easy_) access to titles.

Leaving the utils in there since the kasten is still using them.